+++ src/app.controller.ts
@@ -1,6 +1,6 @@
 import { Request, Response } from '@app/core';
 import { Controller, Get, Req, Res } from '@nestjs/common';
-import { AppService } from './app.service';
+import { AppService } from '@app/app.service';
 import AppLogger from '@app/core/logger/AppLogger';
 @Controller()
 export class AppController {
