+++ src/app.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
+import { AppController } from '@app/app.controller';
+import { AppService } from '@app/app.service';
 import { CoreModule } from '@app/core/module';
 import { DatabaseModule } from '@app/db/database.module';
 import { UsersModule } from '@app/components/users/users.module';
 
-import { ThrottleModule } from '@app/core/rate limiter/throttle.module';
+import { ThrottleModule } from '@app/core/rate-limiter/throttle.module';
 
 @Module({
   imports: [CoreModule, DatabaseModule, UsersModule, ThrottleModule],
