+++ src/components/users/repository/db/user.repository.ts
@@ -1,7 +1,7 @@
-import { CreateUserDto } from '../../dto/create.user.dto';
-import { UpdateUserDto } from '../../dto/update.user.dto';
-import { User } from '../../entities/user.entity';
-import { UserRepository } from '../user.repository';
+import { CreateUserDto } from '@app/components/users/dto/create.user.dto';
+import { UpdateUserDto } from '@app/components/users/dto/update.user.dto';
+import { User } from '@app/components/users/entities/user.entity';
+import { UserRepository } from '@app/components/users/repository/user.repository';
 import { EntityRepository, Repository, UpdateResult } from 'typeorm';
 import { Injectable } from '@nestjs/common';
 @Injectable()
