+++ src/components/users/repository/user.repository.ts
@@ -1,7 +1,7 @@
 import { UpdateResult } from 'typeorm';
-import { CreateUserDto } from '../dto/create.user.dto';
-import { UpdateUserDto } from '../dto/update.user.dto';
-import { User } from '../entities/user.entity';
+import { CreateUserDto } from '@app/components/users/dto/create.user.dto';
+import { UpdateUserDto } from '@app/components/users/dto/update.user.dto';
+import { User } from '@app/components/users/entities/user.entity';
 
 export interface UserRepository {
   findUser(id: string): Promise<User>;
