+++ src/components/users/services/users.service.ts
@@ -1,10 +1,10 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { CreateUserDto } from '../dto/create.user.dto';
-import { UpdateUserDto } from '../dto/update.user.dto';
-import { User } from '../entities/user.entity';
-import { UserDbRepository } from '../repository/db/user.repository';
-import { UserRepository } from '../repository/user.repository';
+import { CreateUserDto } from '@app/components/users/dto/create.user.dto';
+import { UpdateUserDto } from '@app/components/users/dto/update.user.dto';
+import { User } from '@app/components/users/entities/user.entity';
+import { UserDbRepository } from '@app/components/users/repository/db/user.repository';
+import { UserRepository } from '@app/components/users/repository/user.repository';
 
 @Injectable()
 export class UsersService {
