+++ src/components/users/users.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { UsersService } from './services/users.service';
-import { UsersController } from './users.controller';
-import { UserDbRepository } from './repository/db/user.repository';
+import { UsersService } from '@app/components/users/services/users.service';
+import { UsersController } from '@app/components/users/users.controller';
+import { UserDbRepository } from '@app/components/users/repository/db/user.repository';
 
 @Module({
   imports: [TypeOrmModule.forFeature([UserDbRepository])],
