+++ src/config/index.ts
@@ -1,8 +1,8 @@
-import app from './app';
-import db from './database';
-import settings from './settings';
-import services from './services';
-import crypto from './crypto';
-import throttler from './throttler';
+import app from '@app/config/app';
+import db from '@app/config/database';
+import settings from '@app/config/settings';
+import services from '@app/config/services';
+import crypto from '@app/config/crypto';
+import throttler from '@app/config/throttler';
 
 export default [app, db, crypto, settings, throttler, services];
