+++ test/e2e/user.e2e.ts
@@ -3,8 +3,8 @@ import { INestApplication } from '@nestjs/common';
 import * as request from 'supertest';
 import { AppModule } from '@app/app.module';
 import coreBootstrap from '@app/core/bootstrap';
-import { userStub } from '../mock/user.stub';
-import { updateUserStub } from '../mock/user.update.stub';
+import { userStub } from '@test/mock/user.stub';
+import { updateUserStub } from '@test/mock/user.update.stub';
 import { setupAPIVersioning } from '@app/core/api.versioning';
 import { setupSwagger } from '@app/swagger';
 import { StatusCodes } from 'http-status-codes';
