+++ test/unit/config/api-versioning-test.spec.ts
@@ -3,8 +3,8 @@ import { INestApplication } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import * as request from 'supertest';
 import coreBootstrap from '@app/core/bootstrap';
-import appConfig1 from '../core/guards/config/appConfig1';
-import appConfig2 from '../core/guards/config/appConfig2';
+import appConfig1 from '@test/unit/core/guards/config/appConfig1';
+import appConfig2 from '@test/unit/core/guards/config/appConfig2';
 import { DatabaseModule } from '@app/db/database.module';
 import { UsersModule } from '@app/components/users/users.module';
 import { setupAPIVersioning } from '@app/core/api.versioning';
